pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "qlik-sense-mcp-server"
version = "1.4.0"
description = "MCP Server for Qlik Sense Enterprise APIs"
readme = "README.md"
license = {text = "MIT"}
authors = [{name = "Stanislav Chernov", email = "bintocher@yandex.com"}]
keywords = ["mcp", "qlik", "qlik-sense", "analytics", "model-context-protocol"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Information Analysis"
]
dependencies = [
    "mcp>=1.1.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "websocket-client>=1.6.0",
    "starlette>=0.37.0",
    "uvicorn>=0.30.0"
]
requires-python = ">=3.12"

[project.optional-dependencies]
dev = [
    "build>=0.10.0",
    "twine>=4.0.0",
    "bump2version>=1.0.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0"
]

[project.urls]
Homepage = "https://github.com/bintocher/qlik-sense-mcp"
Repository = "https://github.com/bintocher/qlik-sense-mcp"
Issues = "https://github.com/bintocher/qlik-sense-mcp/issues"

[project.scripts]
qlik-sense-mcp-server = "qlik_sense_mcp_server.server:main"
qlik-sense-mcp-gateway = "qlik_sense_mcp_server.remote_gateway:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
